refactor for react-router, login functionality
