+++ client/index.html
@@ -12,4 +12,3 @@
   </header>
 </body>
 </html>
-
