+++ package.json
@@ -15,6 +15,8 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^1.5.0",
+    "bcryptjs": "^2.4.3",
+    "brcypt": "^1.0.1",
     "cors": "^2.8.5",
     "eslint-config-airbnb": "^19.0.4",
     "express": "^4.18.2",
@@ -48,15 +50,3 @@
     "webpack-dev-server": "^4.15.1"
   }
 }
-
-
-
-
-
-
-
-
-
-
-
-
