+++ server/controllers/userController.js
@@ -1,6 +1,75 @@
 const User = require('../models/userModel');
-
 const userController = {};
+const bcrypt = require('bcryptjs');
+
+//sign up
+userController.signUp = (req, res, next) =>{
+  const { username, password } = req.body;
+  console.log(req.body);
+  res.locals.success = true;
+  if(!username || !password) {
+    res.locals.success = false;
+    return next({
+      log: 'username or password is null, signup failed',
+      message: 'user name or password is null, sign up failed'
+    });
+  }
+  if (res.locals.success) {
+    User.create({ username, password })
+      .then(data => {
+        console.log(data);
+        res.locals.user = data;
+        res.locals.success = true;
+        console.log(`User ${username} successfully created!`);
+        return next();
+      })
+      .catch(err => {
+        res.locals.success = false;
+        return next({
+          err,
+          message: 'error creating account in userController.signUp',
+          log: 'error creating account in userController.signUp'
+        });
+      });
+  }
+};
+
+//login
+userController.login = (req, res, next) => {
+  const { username, password } = req.body;
+  if(!username || !password) {
+    return next({
+      log: 'username or password is null, login failed',
+      message: 'user name or password is null, login failed'
+    });
+  }
+  User.findOne({ username })
+    .then(data => {
+      if(data) {
+        res.locals.user = data;
+        bcrypt
+          .compare(password, res.locals.user.password)
+          .then(result => {
+            if (result) res.locals.success = true;
+            console.log('locals success', res.locals.success)
+            return next();
+          })
+          .catch((err)=>{
+            return next({
+              log: 'userController.login error, bcrypt',
+              message: 'userController.login error, bcrypt'
+            });
+          });
+      } else return next();
+    })
+    .catch(err => {
+      return next({
+        err,
+        message: 'login failed: error in userController.login',
+        log: 'login failed: error in userController.login'
+      });
+    });
+}
 
 // add a user
 userController.addUser = (req, res, next) => {
